+++ packages/core/CHANGELOG.md
@@ -11,6 +11,7 @@
     * Change the `referenceToSet` function into a proper symbol constant.
       *Note* that this is technically a breaking change, but this function should essentially only be used internally.
     * Add a `isReferenceToSet` function which satisfies `isReferenceToSet(referenceToSet)`.
+    * Add a `tryToRenderAsText` convenience function to nicely render a reference as text.
 * Deserialization now reports on serialized nodes that can’t be deserialized, and doesn’t add/set `null` anymore.
 * Serialization skips references that are unresolved and for which no `resolveInfo` can be obtained.
 * *Narrow* the type of the argument of the `asIds` function to `Node[]`.