+++ packages/utilities/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Deprecate the `readFileAsJson` and `writeJsonAsFile` functions in favor of the `readFileAsJsonSync` and `writeJsonAsFileSync` functions from `@lionweb/node-utils`.
   (This change has been propagated throughout the entire codebase.)
+* W.r.t. the `sortedSerializationChunk` function:
+  * Give it an additional, optional flag argument `sortConnections` that determines whether containments, references, and annotations are sorted by ID.
+  * It now sorts used languages and features by full meta-pointer, not just its key (language’s key for used language, feature’s key for feature).
 
 
 ## 0.9.0